hooked react-redux into allProduct
