+++ client/components/AllProduct.js
@@ -1,48 +1,64 @@
-import React from 'react'
-import {Link} from 'react-router-dom'
+import React, { Component } from 'react';
 import { connect } from 'react-redux'
-// import { withRouter }
+
+import store from '../store'
+import { fetchProducts } from '../store/product'
+import { fetchOrder } from '../store/order'
+import { Link} from 'react-router-dom'
 
 
-const fakeProducts = [
-{ name: 'tuner and hooch', category: 'comedy', description: 'andfgimals tom hanks + dog', photos: 'https://i.ytimg.com/vi/kEJ0VY9jW9E/maxresdefault.jpg', stock: 28, price: 70 },
-{ name: 'tuner and sandwhic', category: 'centralfds', description: 'anadfimals tom hanks + dog', photos: 'https://frontrowmovies.net/Actionmovies.jpg', stock: 13, price: 90 },
-{ name: 'tuner and bread', category: 'central', description: 'anidmals tom hanks + dog', photos: 'http://img.wennermedia.com/social/rs-180704-AM.jpg', stock: 68, price: 40 },
-];
+class AllProduct extends Component {
+  constructor(props) {
+    super(props)
+  }
 
-function AllProduct(props){
-  console.log("THIS IS PROPS", props);
-  return (
-    <div>
-      <h1>All Product List!!!!!!</h1>
-      <Link to="/review">Review</Link>
-      <hr />
-      {
-        fakeProducts.map(product => {
-          return (
-           <div key={product.name}>
-             <h1>{product.name}</h1>
-             <img src={product.photos} />
-             <Link to="/single-product">View </Link> |
-             <Link to="/cart"> Add to Cart </Link>  |
-           </div>
-        )
-        })
+  componentDidMount() {
+    const productsThunk = fetchProducts()
+    store.dispatch(productsThunk)
+    const orderThunk = fetchOrder(1) // hard coding an orderID
+    store.dispatch(orderThunk)
+  }
 
-      }
 
-    </div>
-  )
+  render() {
+    console.log(this.props, 'the props')
+
+    return (
+      <div>
+        <h1>All Product List!!!!!!</h1>
+        <Link to="/review">Review</Link>
+        <hr />
+        {
+          this.props.product.map(product => {
+            return (
+             <div key={product.name}>
+               <h1>{product.name}</h1>
+               <img src={product.photos} />
+               <Link to="/single-product">View </Link> |
+               <Link to="/cart"> Add to Cart </Link>  |
+              </div>
+          )
+          })
+
+        }
+
+      </div>
+    )
+  }
 }
 
 
+// CONTAINER
+
 const mapStateToProps = (state) => {
   return {
-    allProducts: state.allProducts,
+    product: state.product,
     order: state.order
   }
 }
 
+// call connect function from react-redux, pass it mapState, and invoke with the presentational component (this component itself)
 export default connect(mapStateToProps)(AllProduct)
-//export default AllProduct;
+// don't need map dispatch yet because no methods being called
 
+//PROP TYPES
