+++ client/components/main.js
@@ -13,7 +13,7 @@ import Test from './test'
  */
 const Main = (props) => {
   const {children, handleClick, isLoggedIn} = props
-
+// console.log(props, 'in main')
   return (
     <div>
       <h1>Jean-Claude-Schwarzenegger-INC</h1>
@@ -36,7 +36,7 @@ const Main = (props) => {
       </nav>
       <hr />
       {children}
-      <Test />
+        <Test />
     </div>
   )
 }
@@ -46,7 +46,7 @@ const Main = (props) => {
  */
 const mapState = (state) => {
   return {
-    isLoggedIn: !!state.user.id
+    isLoggedIn: !!state.user.id,
   }
 
 }
