+++ client/components/test.js
@@ -1,5 +1,5 @@
 import React from 'react';
-import OrderState from './orderState.js'
+import AllProduct from './allProduct.js'
 
 
 
@@ -7,7 +7,7 @@ export default function testWrapper(){
     return (
       <div>
         <h3>TESTS GO HERE</h3>
-       <OrderState />
+       <AllProduct />
       </div>
     )
