+++ client/routes.js
@@ -4,19 +4,30 @@ import {Router} from 'react-router'
 import {Route, Switch} from 'react-router-dom'
 import PropTypes from 'prop-types'
 import history from './history'
-import {Main, Login, Signup, UserHome, Review, AllProduct, Cart, SingleProduct} from './components'
-import {me} from './store'
+import {Main, Login, Signup, UserHome, Review, Cart, SingleProduct} from './components'
+
+import store, {me} from './store'
+import { fetchOrder } from './store/order'
+import { fetchProducts } from './store/product'
+
 
 /**
  * COMPONENT
  */
 class Routes extends Component {
   componentDidMount () {
+    const productsThunk = fetchProducts()
+    store.dispatch(productsThunk)
+    const orderThunk = fetchOrder(2) // hard coding an orderID
+    store.dispatch(orderThunk)
     this.props.loadInitialData()
   }
 
+
+
   render () {
     const {isLoggedIn} = this.props
+    // console.log(this.props, 'in routes')
 
     return (
       <Router history={history}>
@@ -26,7 +37,6 @@ class Routes extends Component {
             <Route path="/login" component={Login} />
             <Route path="/signup" component={Signup} />
             <Route path="/review" component={Review} />
-            <Route path="/all-products" component={AllProduct} />
             <Route path="/single-product" component={SingleProduct} />
             <Route path="/cart" component={Cart} />
 
@@ -53,7 +63,9 @@ const mapState = (state) => {
   return {
     // Being 'logged in' for our purposes will be defined has having a state.user that has a truthy id.
     // Otherwise, state.user will be an empty object, and state.user.id will be falsey
-    isLoggedIn: !!state.user.id
+    isLoggedIn: !!state.user.id,
+    product: state.product,
+    order: state.order
   }
 }
